fix(sync): schedule_hours i64 → i32 to match INT4 column
SyncJob.schedule_hours was declared as i64 (INT8) but sync_jobs.schedule_hours
is integer (INT4) — sqlx refused to decode the row on every page load.

+++ src/sync/mod.rs
@@ -20,7 +20,7 @@ pub struct TaskDef {
 #[derive(Debug, Clone, sqlx::FromRow)]
 pub struct SyncJob {
     pub name: String,
-    pub schedule_hours: i64,
+    pub schedule_hours: i32,
     pub last_started_at: Option<OffsetDateTime>,
     pub last_finished_at: Option<OffsetDateTime>,
     pub last_status: Option<String>,
